rearrange files for correct class hierarchy (task #3284)
