--- src/WidgetHandlers/KnobChartWidgetHandler.php
@@ -1,9 +1,0 @@
-<?php
-namespace Search\WidgetHandlers;
-
-use Search\WidgetHandlers\BaseReportWidget;
-
-class KnobChartWidgetHandler extends BaseReportWidget
-{
-
-}
