--- src/WidgetHandlers/WidgetReportInterface.php
@@ -1,7 +1,0 @@
-<?php
-namespace Search\WidgetHandlers;
-
-interface WidgetReportInterface
-{
-    public function validate(array $data = []);
-}
